AuditLog Class

This class is used internally to store audit logs for each request.

Methods

Public methods

AuditLog$new()

AuditLog$disable()

AuditLog$is_enabled()

AuditLog$set_request_body()

AuditLog$set_response_body()

AuditLog$set_ip_address()

AuditLog$set_user_agent()

AuditLog$set_event_type()

AuditLog$set_study_id()

AuditLog$set_response_code()

AuditLog$validate_log()

AuditLog$persist()

AuditLog$clone()

Method new()

Usage

AuditLog$new(request_method, endpoint_url)

Method disable()

Usage

AuditLog$disable()

Method is_enabled()

Usage

AuditLog$is_enabled()

Method set_request_body()

Usage

AuditLog$set_request_body(request_body)

Method set_response_body()

Usage

AuditLog$set_response_body(response_body)

Method set_ip_address()

Usage

AuditLog$set_ip_address(ip_address)

Method set_user_agent()

Usage

AuditLog$set_user_agent(user_agent)

Method set_event_type()

Usage

AuditLog$set_event_type(event_type)

Method set_study_id()

Usage

AuditLog$set_study_id(study_id)

Method set_response_code()

Usage

AuditLog$set_response_code(response_code)

Method validate_log()

Usage

AuditLog$validate_log()

Method persist()

Usage

AuditLog$persist()

Method clone()

The objects of this class are cloneable with this method.

Usage

AuditLog$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.